Rename Trema::Logger -> Trema::DefaultLogger (refs #264).
Formerly, if a programmer tries to use Trema combined with ActiveSupport library, trema run failed with "Logger is not a class" error. This is because ActiveSupport defines Logger class at the top level, and trema run includes Trema module at start-up, therefore a type-mismatch between the definition of Trema::Logger and AR's Logger occurred.

This fix renames Trema::Logger as Trema::DefaultLogger to avoid the name conflict.