fix loss of some OFPT_PORT_MOD messages
The ofp_port_mod.advertise field is a bitmask of OFPPF_* constants and
its validation must be based on the PORT_FEATURES constant (as it is in
the validate_phy_port() function). But the validate_port_mod() function
used PORT_CONFIG constant instead and erroneously failed the check for
a correct message with any of the following bits set:
* OFPPF_COPPER
* OFPPF_FIBER
* OFPPF_AUTONEG
* OFPPF_PAUSE
* OFPPF_PAUSE_SYM

The controller would log an error with code ERROR_INVALID_PORT_FEATURES
(-49) and discard the message. This change fixes the bug.