add -r/--no_flow_cleanup option to "trema run"
The OF 1.0 specification leaves it up to implementation whether the
switch flushes its flow table(s) upon controller disconnection. Some
implementations retain the flows and make it possible to design hitless
restart applications on base of that.

The switch handler process of Trema would delete all flows from a
switch during connection setup unless run with the --no-flow-cleanup
option. This is possible through the SwitchManager.no_flow_cleanup
accessor, which the "trema run" use case used not to manage.

There is now the new command-line option to "trema run" to change this
default behaviour.